Fix and improve build scripts

+++ scripts/build-package.sh
@@ -12,6 +12,9 @@ while [[ "$#" -gt 0 ]]; do
     -unix)
     use_unix=yes
     ;;
+    -reuse)
+    reuse_build=yes
+    ;;
     -*)
     echo "error: unknown option \"$1\""
     exit 1
@@ -23,18 +26,35 @@ while [[ "$#" -gt 0 ]]; do
   shift
 done
 
+get_platform_name() {
+  if [[ "$OSTYPE" == "msys" ]]; then
+    echo "windows"
+  elif [[ "$OSTYPE" == "darwin"* ]]; then
+    echo "macos"
+  elif [[ "$OSTYPE" == "linux"* || "$OSTYPE" == "freebsd"* ]]; then
+    echo "linux"
+  else
+    echo "UNSUPPORTED-OS"
+  fi
+}
+
 ext=""
+arch="$(uname -m)"
+platform="$(get_platform_name)"
 if [[ "$OSTYPE" == "msys"* || "$OSTYPE" == "mingw"* ]]; then
   ext=".exe"
   strip="strip --strip-all"
 else
   strip=strip
+  os=linux
 fi
 
 builddir=".build-release"
 
-rm -fr "$builddir"
-meson setup --buildtype=release $blas_option "$builddir"
+if [ -z ${reuse_build+x} ]; then
+  rm -fr "$builddir"
+  meson setup --buildtype=release $blas_option "$builddir"
+fi
 ninja -C "$builddir"
 
 rundir="packages/gsl-shell"
@@ -54,4 +74,6 @@ cp -r data/. "$datadir"
 cp resources/gsl-shell.desktop resources/gsl-shell.svg "$rundir"
 
 $strip "$bindir/gsl-shell$ext" "$bindir/gsl-shell-gui$ext"
+
+tar czf gsl-shell-$platform-$arch.tar.gz -C "$rundir/.." gsl-shell
